Refactor: RP-109 대대적인 auth 아키텍처 수정
