+++ src/main/java/com/cherrypick/backend/domain/auth/application/Oauth2Client.java
@@ -1,0 +1,4 @@
+package com.cherrypick.backend.domain.auth.application;
+
+public class Oauth2Client {
+}
