+++ src/main/java/com/cherrypick/backend/domain/auth/domain/vo/UserEnv.java
@@ -1,0 +1,4 @@
+package com.cherrypick.backend.domain.auth.domain.vo;
+
+public class UserEnv {
+}
