+++ src/main/java/com/cherrypick/backend/domain/auth/domain/vo/token/AccessToken.java
@@ -1,0 +1,4 @@
+package com.cherrypick.backend.domain.auth.domain.vo.token;
+
+public class AccessToken {
+}
