+++ src/main/java/com/cherrypick/backend/domain/auth/infra/jwt/AccessTokenProvider.java
@@ -1,0 +1,4 @@
+package com.cherrypick.backend.domain.auth.infra.jwt;
+
+public class AccessTokenProvider {
+}
