+++ src/main/java/com/cherrypick/backend/domain/auth/infra/store/RegisterTokenStore.java
@@ -1,0 +1,4 @@
+package com.cherrypick.backend.domain.auth.infra.store;
+
+public class RegisterTokenStore {
+}
