+++ src/main/java/com/cherrypick/backend/domain/auth/presentation/dto/AuthRequestDTOs.java
@@ -1,4 +1,4 @@
-package com.cherrypick.backend.domain.oauth.dto;
+package com.cherrypick.backend.domain.auth.presentation.dto;
 
 public class AuthRequestDTOs
 {
