+++ src/main/java/com/cherrypick/backend/global/util/AuthUtil.java
@@ -1,21 +1,17 @@
 package com.cherrypick.backend.global.util;
 
-import com.cherrypick.backend.domain.user.dto.AuthenticationDetailDTO;
+import com.cherrypick.backend.domain.auth.domain.vo.AuthenticatedUser;
 import com.cherrypick.backend.domain.user.enums.Role;
 import com.cherrypick.backend.global.exception.BaseException;
 import com.cherrypick.backend.global.exception.enums.UserErrorCode;
-import jakarta.transaction.NotSupportedException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 
-import java.util.Collection;
 import java.util.List;
-import java.util.Set;
 
 @Slf4j
 public class AuthUtil
@@ -29,11 +25,11 @@ public class AuthUtil
         if (authentication == null || !authentication.isAuthenticated()) return false;
 
         var principal = authentication.getPrincipal();
-        return principal instanceof AuthenticationDetailDTO;
+        return principal instanceof AuthenticatedUser;
     }
 
 
-    public static AuthenticationDetailDTO getUserDetail() {
+    public static AuthenticatedUser getUserDetail() {
         var authentication = SecurityContextHolder.getContext().getAuthentication();
 
         if (authentication == null || !authentication.isAuthenticated()) {
@@ -42,11 +38,11 @@ public class AuthUtil
 
         var principal = authentication.getPrincipal();
 
-        if (!(principal instanceof AuthenticationDetailDTO)) {
+        if (!(principal instanceof AuthenticatedUser)) {
             throw new BaseException(UserErrorCode.SECURITY_AUTHENTICATION_REQUIRED);
         }
 
-        return (AuthenticationDetailDTO) principal;
+        return (AuthenticatedUser) principal;
     }
 
     // 유저의 권한을 전환.
@@ -55,7 +51,7 @@ public class AuthUtil
 
         Authentication currentAuth = SecurityContextHolder.getContext().getAuthentication();
         // 기존의 저장된 유저 정보
-        AuthenticationDetailDTO principal = (AuthenticationDetailDTO)currentAuth.getPrincipal();
+        AuthenticatedUser principal = (AuthenticatedUser)currentAuth.getPrincipal();
 
 
         List<GrantedAuthority> updatedAuthorities = List.of(
@@ -63,7 +59,7 @@ public class AuthUtil
         );
 
         // 새로운 유저 정보
-        var newPrincipal = AuthenticationDetailDTO.builder()
+        var newPrincipal = AuthenticatedUser.builder()
                 .userId(principal.userId())
                 .nickname(principal.nickname())
                 .role(role)
